incorporated PR comments
