+++ tools/c7n_gcp/c7n_gcp/filters/__init__.py
@@ -5,4 +5,3 @@ from .metrics import GCPMetricsFilter
 from .sccfindings import SecurityComandCenterFindingsFilter
 from .iampolicy import IamPolicyFilter
 from .alerts import AlertsFilter
-from .essential_contacts import EssentialContactsFilter
